game winning modal added
The game cannot be "won" at this point, there currently is a bug with
determining the game winner in the java code. Once that is fixed this
commit will make sense.

Commented out the lines which should do this operation in the
StoredProcedures. Once the bug is fixed in Game.java then these lines
should be reinstated.